requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
linkedin-api>=2.0.0
tenacity>=8.2.0
newsapi-python>=0.2.7
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0 